Move timed_async to tick
